requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
pydantic>=2.5.0
pytest>=7.4.0
